Refactor: Move artery input to physics
