+++ src/art_net/4C_art_net_artery_evaluate.cpp
@@ -8,9 +8,9 @@
 #include "4C_art_net_artery.hpp"
 #include "4C_art_net_artery_ele_action.hpp"
 #include "4C_art_net_artery_ele_factory.hpp"
+#include "4C_art_net_input.hpp"
 #include "4C_fem_discretization.hpp"
 #include "4C_fem_general_utils_fem_shapefunctions.hpp"
-#include "4C_inpar_bio.hpp"
 #include "4C_linalg_utils_sparse_algebra_math.hpp"
 #include "4C_mat_cnst_1d_art.hpp"
 #include "4C_utils_enum.hpp"
